+++ frontend/src/pages/CollectionsPage.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import CollectionsList from '../components/collections/CollectionsList'
-
-const CollectionsPage: React.FC = () => {
-  return <CollectionsList />
-}
-
-export default CollectionsPage