+++ frontend/src/pages/ProjectsPage.tsx
@@ -1,8 +1,0 @@
-import React from 'react'
-import ProjectsList from '../components/projects/ProjectsList'
-
-const ProjectsPage: React.FC = () => {
-  return <ProjectsList />
-}
-
-export default ProjectsPage